+++ lib/binance_client.rb
@@ -10,6 +10,7 @@ require "binance_client/client"
 require "binance_client/models/base_model"
 require "binance_client/models/order_book_entry"
 require "binance_client/models/order_book"
+require "binance_client/models/asset_balance"
 
 require "binance_client/requests/base_request"
 require "binance_client/responses/base_response"
@@ -18,11 +19,13 @@ require "binance_client/requests/account_snapshot_request"
 require "binance_client/requests/get_all_request"
 require "binance_client/requests/book_ticker_request"
 require "binance_client/requests/order_book_depth_request"
+require "binance_client/requests/sub_account_assets_request"
 require "binance_client/responses/system_status_response"
 require "binance_client/responses/account_snapshot_response"
 require "binance_client/responses/get_all_response"
 require "binance_client/responses/book_ticker_response"
 require "binance_client/responses/order_book_depth_response"
+require "binance_client/responses/sub_account_assets_response"
 
 module BinanceClient
   class Error < StandardError; end