+++ lib/binance_client/responses/sub_account_assets_response.rb
@@ -1,0 +1,16 @@
+module BinanceClient
+  class SubAccountAssetsResponse < BaseResponse
+
+    attribute(:balances, Array[BinanceClient::AssetBalance], {
+      lazy: true,
+      default: :default_balances,
+    })
+
+    def default_balances
+      body["balances"].map do |balance_hash|
+        AssetBalance.new(**balance_hash.symbolize_keys)
+      end
+    end
+
+  end
+end