+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Allow easy access to the `used_weight` header in a response by calling `#used_weight(interval)`
+- `#sub_account_assets` to get the assets of a sub-account
 
 ## [1.1.0]
 ### Added